Fix #2 - refactor hokuyo to proper package name
