+++ hokuyo/src/main/java/com/github/ros_java/rosjava_extras/hokuyo/LaserScan.java
@@ -1,12 +1,12 @@
 /*
  * Copyright (C) 2011 Google Inc.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
  * the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
@@ -14,13 +14,13 @@
  * the License.
  */
 
-package org.ros.android.hokuyo;
+package com.github.ros_java.rosjava_extras.hokuyo;
 
 import org.ros.message.Time;
 
 /**
  * Represents a collection of range reading from the sensor.
- * 
+ *
  * @author moesenle@google.com (Lorenz Moesenlechner)
  * @author damonkohler@google.com (Damon Kohler)
  */
