+++ hokuyo/src/main/java/com/github/ros_java/rosjava_extras/hokuyo/LaserScanPublisher.java
@@ -1,12 +1,12 @@
 /*
  * Copyright (C) 2011 Google Inc.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
  * the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package org.ros.android.hokuyo;
+package com.github.ros_java.rosjava_extras.hokuyo;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
@@ -79,14 +79,14 @@ public class LaserScanPublisher extends AbstractNodeMain {
   /**
    * Construct a LaserScan message from sensor readings and the laser
    * configuration.
-   * 
+   *
    * Also gets rid of readings that don't contain any information.
-   * 
+   *
    * Some laser scanners have blind areas before and after the actual detection
    * range. These are indicated by the frontStep and the lastStep properties of
    * the laser's configuration. Since the blind values never change, we can just
    * ignore them when copying the range readings.
-   * 
+   *
    * @param laserFrame
    *          the laser's sensor frame
    * @param scan
