+++ hokuyo/src/main/java/com/github/ros_java/rosjava_extras/hokuyo/LaserScannerConfiguration.java
@@ -1,12 +1,12 @@
 /*
  * Copyright (C) 2011 Google Inc.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
  * the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package org.ros.android.hokuyo;
+package com.github.ros_java.rosjava_extras.hokuyo;
 
 public interface LaserScannerConfiguration {
 
@@ -42,7 +42,7 @@ public interface LaserScannerConfiguration {
    * Returns the first meaningful range reading. The laser might have a blind
    * area at the beginning of the scan range. Range readings are generated for
    * this area, they do not contain any useful information though.
-   * 
+   *
    * @return The index of the first meaningful range reading.
    */
   int getFirstStep();
@@ -51,7 +51,7 @@ public interface LaserScannerConfiguration {
    * Returns the last meaningful range reading. The laser might have a blind
    * area at the end of the scan range. Range readings are generated for this
    * area, they do not contain any useful information though.
-   * 
+   *
    * @return The index of the last meaningful range reading.
    */
   int getLastStep();
@@ -59,7 +59,7 @@ public interface LaserScannerConfiguration {
   /**
    * Returns the front step of the laser. The front step is the index of the
    * reading that is pointing directly forward.
-   * 
+   *
    * @return The index of the front step.
    */
   int getFrontStep();
