+++ hokuyo/src/main/java/com/github/ros_java/rosjava_extras/hokuyo/scip20/Configuration.java
@@ -1,12 +1,12 @@
 /*
  * Copyright (C) 2011 Google Inc.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
  * the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
@@ -14,12 +14,12 @@
  * the License.
  */
 
-package org.ros.android.hokuyo.scip20;
+package com.github.ros_java.rosjava_extras.hokuyo.scip20;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 
-import org.ros.android.hokuyo.LaserScannerConfiguration;
+import com.github.ros_java.rosjava_extras.hokuyo.LaserScannerConfiguration;
 
 
 
@@ -138,7 +138,7 @@ public class Configuration implements LaserScannerConfiguration {
    * Returns the first meaningful range reading. The laser might have a blind
    * area at the beginning of the scan range. Range readings are generated for
    * this area, they do not contain any useful information though.
-   * 
+   *
    * @return the index of the first meaningful range reading
    */
   @Override
@@ -150,7 +150,7 @@ public class Configuration implements LaserScannerConfiguration {
    * Returns the last meaningful range reading. The laser might have a blind
    * area at the end of the scan range. Range readings are generated for this
    * area, they do not contain any useful information though.
-   * 
+   *
    * @return the index of the last meaningful range reading
    */
   @Override
@@ -161,7 +161,7 @@ public class Configuration implements LaserScannerConfiguration {
   /**
    * Returns the front step of the laser. The front step is the index of the
    * reading that is pointing directly forward.
-   * 
+   *
    * @return the index of the front step
    */
   @Override
