+++ hokuyo/src/test/java/com/github/ros_java/rosjava_extras/hokuyo/FakeLaserScannerConfiguration.java
@@ -1,12 +1,12 @@
 /*
  * Copyright (C) 2011 Google Inc.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
  * the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
@@ -14,9 +14,7 @@
  * the License.
  */
 
-package org.ros.android.hokuyo;
-
-import org.ros.android.hokuyo.LaserScannerConfiguration;
+package com.github.ros_java.rosjava_extras.hokuyo;
 
 /**
  * @author moesenle@google.com (Lorenz Moesenlechner)
