+++ hokuyo/src/test/java/com/github/ros_java/rosjava_extras/hokuyo/scip20/ConfigurationTest.java
@@ -1,12 +1,12 @@
 /*
  * Copyright (C) 2011 Google Inc.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
  * the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
@@ -14,9 +14,7 @@
  * the License.
  */
 
-package org.ros.android.hokuyo.scip20;
-
-import org.ros.android.hokuyo.scip20.Configuration;
+package com.github.ros_java.rosjava_extras.hokuyo.scip20;
 
 import junit.framework.TestCase;
 
@@ -24,7 +22,7 @@ import junit.framework.TestCase;
  * @author damonkohler@google.com (Damon Kohler)
  */
 public class ConfigurationTest extends TestCase {
-  
+
   private Configuration.Builder builder;
 
   @Override
@@ -36,12 +34,12 @@ public class ConfigurationTest extends TestCase {
     builder.parseModel("MODL:URG-04LX(Hokuyo Automatic Co., Ltd.);");
     assertEquals("URG-04LX(Hokuyo Automatic Co., Ltd.)", builder.build().getModel());
   }
-  
+
   public void testParseIntegerValue() {
     // NOTE(damonkohler): We leave off the trailing ";" here because it is
     // stripped before parsing.
     assertEquals(20, builder.parseIntegerValue("DMIN", "DMIN:20"));
   }
-  
- 
+
+
 }
